use incoming block_size as block upload max
Prior to this commit, a hardcoded value of 1024 was used as the maximum
block upload size. This commit uses the `block_size` value passed in to the
callback which will be set to the max value by the Golioth Firmware SDK.
This value is then reset to the actual number of byte written by the
callback, which should always be max except for the final block.

Signed-off-by: Mike Szczys <mike@golioth.io>